Enable autoCloseBrackets in CodeMirror editor
See http://codemirror.net/doc/manual.html#addon_closebrackets for a description. Among other things, this makes it much easier to type the curly braces of functions when they have a new line between them – you can just type ‘{’, ‘Enter’.

I also added a comment naming the source of the Tab key map. This explains to readers why that particular implementation for Tab was chosen, instead of a more sophisticated one that indents selected lines when there is a selection.