Make CodeMirror reindent lines on paste
The project from which this code is adapted, brackets-paste-and-indent, uses the Unlicense, so it is fine to take and adapt code from it.

My changes/adaptations were combining the code into one block so that it is more self-contained, and reformatting its style to match the rest of this project.