Realism: Passengers often don't press buttons already pressed
In Real Life, a passenger may press the same button many times.
Also, many passengers won't press the button for the floor they
want if it's already lit. So elevator functions should not be
able to reliably count the number of times a given floor's button
was pressed. Use the same pattern as for the floors -- check for
a change in value and only emit the event(s) if the value changed.