Enable replaying a challenge sequence rather than resetting
Restarting a challenge creates a new random sequence of user arrivals
and floor requests. This means it's possible to pass a given challenge
simply by retrying after each failure.

Instead, save the sequence of user arrivals and floor requests as they
are generated. Replay them when we're restarting the same challenge after
a failure. Navigation to a different challenge or restarting after success
generates new sequences as before.

Note that saving the sequence of user arrivals and floor requests does
not perfectly save the failed game for replay since the elevators
users board, X positions, and the slots users get assigned are
still randomized. While those attributes can be saved this constitutes
a reasonable subset that is easy to restore and preserves much of the
difficulty presented by a given arrival and request sequence.