Prepare to Show Individual Wait and Travel Times: Remove stats template
Showing individual wait and travel times is tricky since showing all
three time values at once will result in an even-more cluttered display.
So we will add code to allow the user to select which time to display
on the individual users by clicking on the appropriate row in the
time statistics table.

We must stop re-rendering the stats template on each stats change
because it interferes with receiving input in the stats area. Replace
the templatized rendering with jQuery code searching on specific
element identifiers.